change PageCache to service provider
