+++ web/concrete/src/Cache/CacheServiceProvider.php
@@ -15,6 +15,6 @@ class CacheServiceProvider extends ServiceProvider
         $this->app->singleton('cache', '\Concrete\Core\Cache\Level\ObjectCache');
         $this->app->singleton('cache/request', '\Concrete\Core\Cache\Level\RequestCache');
         $this->app->singleton('cache/expensive', '\Concrete\Core\Cache\Level\ExpensiveCache');
-        $this->app->singleton('cache/page', '\Concrete\Core\Cache\Level\PageCache');
+        $this->app->singleton('cache/page', '\PageCache');
     }
 }
