+++ web/concrete/src/Cache/Level/ObjectCache.php
@@ -1,12 +1,9 @@
 <?php
-
 namespace Concrete\Core\Cache\Level;
 
-
-use Concrete\Core\Cache\Cache;
 use Stash\Pool;
 
-class ObjectCache extends Cache
+class ObjectCache extends \Cache
 {
     protected function init()
     {
