+++ web/concrete/src/Cache/Page/PageCacheRecord.php
@@ -11,7 +11,7 @@ class PageCacheRecord {
     public $cacheRecordKey;
 
 	public function __construct(Page $c, $content, $lifetime) {
-		$cache = PageCache::getLibrary();
+		$cache = \PageCache::getLibrary();
 		$this->setCacheRecordLifetime($lifetime);
 		$this->setCacheRecordKey($cache->getCacheKey($c));
 		$this->setCacheRecordHeaders($cache->getCacheHeaders($c));
@@ -57,7 +57,7 @@ class PageCacheRecord {
 			return true;
 		} else {
 			// invalidate and kill this record.
-			$cache = PageCache::getLibrary();
+			$cache = \PageCache::getLibrary();
 			$cache->purgeByRecord($this);
 		}
 	}
